refactor: migrate files page to servers-component structure
