+++ apps/frontend/src/components/ui/servers/files/FileItem.vue
