+++ apps/frontend/src/components/ui/servers/files/FilesImageViewer.vue
