+++ apps/frontend/src/components/ui/servers/files/FilesNavbar.vue
