+++ apps/frontend/src/components/ui/servers/files/ModalDeleteItem.vue
